Combine auto-setup and server into a single multi-stage dockerfile
`docker buildx bake` doesn't seem to let you override the entrypoint of
the current build context, so when we changed to using a bakefile we
broke the auto-setup image. We can work around this by combining the two
into a single multi-stage dockerfile and picking the target we want to
build separately.

I added two new stages to the server dockerfile to handle this:

1. `server` - the server itself without the auto-setup stuff. This is
equivalent to what the server.Dockerfile would build prior to this
change
2. `auto-setup` - this is the server-with-auto-setup image